Refactor platform specifics in supervisor into sys module
