+++ components/sup/src/lib.rs
@@ -102,6 +102,7 @@ pub mod http_gateway;
 pub mod manager;
 pub mod templating;
 pub mod util;
+mod sys;
 
 use std::env;
 use std::path::PathBuf;
