+++ components/sup/src/util/users.rs
@@ -12,93 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use error::{Result, Error};
 use hcore::os::users;
-use hcore::package::PackageInstall;
+
+pub use sys::users::*;
+use error::{Result, Error};
 
 pub const DEFAULT_USER: &'static str = "hab";
 pub const DEFAULT_GROUP: &'static str = "hab";
-
 static LOGKEY: &'static str = "UR";
-
-/// This function checks to see if a custom SVC_USER and SVC_GROUP has
-/// been specified as part of the package metadata.
-/// If pkg_svc_user and pkg_svc_group have NOT been defined, return None.
-#[cfg(unix)]
-fn get_pkg_user_and_group(pkg_install: &PackageInstall) -> Result<Option<(String, String)>> {
-    let svc_user = try!(pkg_install.svc_user());
-    let svc_group = try!(pkg_install.svc_group());
-    match (svc_user, svc_group) {
-        (Some(user), Some(group)) => Ok(Some((user, group))),
-        _ => {
-            debug!("User/group not specified in package, running with default");
-            Ok(None)
-        }
-    }
-}
-
-/// This function checks to see if a user and group and if:
-///     a) we are root
-///     b) we are the specified user:group
-///     c) fail otherwise
-#[cfg(unix)]
-pub fn assert_pkg_user_and_group(user: String, group: String) -> Result<()> {
-    if let None = users::get_uid_by_name(&user) {
-        return Err(sup_error!(Error::Permissions(format!(
-            "Package requires user {} to \
-                                                            exist, but it doesn't",
-            user
-        ))));
-    }
-    if let None = users::get_gid_by_name(&group) {
-        return Err(sup_error!(Error::Permissions(format!(
-            "Package requires group {} \
-                                                            to exist, but it doesn't",
-            group
-        ))));
-    }
-
-    let current_user = users::get_current_username();
-    let current_group = users::get_current_groupname();
-
-    if let None = current_user {
-        return Err(sup_error!(Error::Permissions(
-            "Can't determine current user".to_string(),
-        )));
-    }
-
-    if let None = current_group {
-        return Err(sup_error!(Error::Permissions(
-            "Can't determine current group".to_string(),
-        )));
-    }
-
-    let current_user = current_user.unwrap();
-    let current_group = current_group.unwrap();
-
-    if current_user == users::root_level_account() {
-        Ok(())
-    } else {
-        if current_user == user && current_group == group {
-            // ok, sup is running as svc_user/svc_group already
-            Ok(())
-        } else {
-            let msg = format!("Package must run as {}:{} or root", &user, &group);
-            return Err(sup_error!(Error::Permissions(msg)));
-        }
-    }
-}
-
-/// Always return Ok (for now) on windows since we just run as the current user
-#[cfg(windows)]
-#[allow(unused)]
-pub fn assert_pkg_user_and_group(user: String, group: String) -> Result<()> {
-    Ok(())
-}
 
 /// checks to see if hab/hab exists, if not, fall back to
 /// current user/group. If that fails, then return an error.
-fn get_default_user_and_group() -> Result<(String, String)> {
+pub fn default_user_and_group() -> Result<(String, String)> {
     let uid = users::get_uid_by_name(DEFAULT_USER);
     let gid = users::get_gid_by_name(DEFAULT_GROUP);
     match (uid, gid) {
@@ -120,28 +45,4 @@ fn get_default_user_and_group() -> Result<(String, String)> {
             }
         }
     }
-}
-
-/// check and see if a user/group is specified in package metadata.
-/// if not, we'll try and use hab/hab.
-/// If hab/hab doesn't exist, try to use (current username, current group).
-/// If that doesn't work, then give up.
-#[cfg(unix)]
-pub fn get_user_and_group(pkg_install: &PackageInstall) -> Result<(String, String)> {
-    if let Some((user, group)) = try!(get_pkg_user_and_group(&pkg_install)) {
-        Ok((user, group))
-    } else {
-        let defaults = try!(get_default_user_and_group());
-        Ok(defaults)
-    }
-}
-
-/// For now we are ignoring any configured user and group
-/// because we do not start the supervisor on windows under
-/// alternate credentials
-#[cfg(windows)]
-#[allow(unused)]
-pub fn get_user_and_group(pkg_install: &PackageInstall) -> Result<(String, String)> {
-    let defaults = try!(get_default_user_and_group());
-    Ok(defaults)
 }
